refactored, added account.js under routes/api
